refactor: modularize pet routes
